+++ packages/create-svelte/templates/default/src/lib/Header.svelte
